+++ packages/components/atoms/f-form-field/CHANGELOG.md
@@ -4,6 +4,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v5.0.0
+-----------------------------
+*Jun 16, 2022*
+
+### Changed
+- Update to `@use` and `@forward` SASS syntax
+
+
 v4.13.1
 -----------------------------
 *Jun 13, 2022*